+++ lib/scenic/primitive/arc.ex
@@ -4,7 +4,34 @@
 #
 
 defmodule Scenic.Primitive.Arc do
-  @moduledoc false
+  @moduledoc """
+  Draw an arc on the screen.
+
+  An arc is a segment that traces part of the outline of a circle. If you are
+  looking for something shaped like a piece of pie, then you want a segment.
+
+  Arcs are often drawn on top of a segment to get an affect where a piece of pie
+  is filled in, but only the curvy edge is stroked.
+
+  Note that you can fill an arc, but that will result in a shape that looks
+  like a potato wedge.
+
+  ## Data
+
+  `{radius, start, finish}`
+
+  The data for an arc is a three-tuple.
+  * `radius` - the radius of the arc
+  * `start` - the starting angle in radians
+  * `finish` - end ending angle in radians
+
+  ## Styles
+
+  This primitive recognizes the following styles
+  * [`hidden`](Scenic.Primitive.Style.Hidden.html) - show or hide the primitive
+  * [`fill`](Scenic.Primitive.Style.Fill.html) - fill in the area of the primitive
+  * [`stroke`](Scenic.Primitive.Style.Stroke.html) - stroke the outline of the primitive. In this case, only the curvy part.
+  """
 
   use Scenic.Primitive
   alias Scenic.Primitive.Sector
@@ -16,6 +43,7 @@ defmodule Scenic.Primitive.Arc do
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be: {radius, start, finish}
@@ -24,6 +52,7 @@ defmodule Scenic.Primitive.Arc do
     """
 
   # --------------------------------------------------------
+  @doc false
   def verify(data) do
     normalize(data)
     {:ok, data}
@@ -32,12 +61,16 @@ defmodule Scenic.Primitive.Arc do
   end
 
   # --------------------------------------------------------
+  @doc false
   @spec normalize({number(), number(), number()}) :: {number(), number(), number()}
   def normalize({radius, start, finish} = data)
       when is_number(start) and is_number(finish) and is_number(radius),
       do: data
 
   # ============================================================================
+  @doc """
+  Returns a list of styles recognized by this primitive.
+  """
   @spec valid_styles() :: [:fill | :hidden | :stroke]
   def valid_styles(), do: @styles