+++ lib/scenic/primitive/rectangle.ex
@@ -4,16 +4,36 @@
 #
 
 defmodule Scenic.Primitive.Rectangle do
-  @moduledoc false
+  @moduledoc """
+  Draw a rectangle on the screen.
+
+  ## Data
+
+  `{width, height}`
+
+  The data for a line is a tuple containing two numbers.
+  * `width` - width of the rectangle
+  * `height` - height of the rectangle
+
+  ## Styles
+
+  This primitive recognizes the following styles
+  * [`hidden`](Scenic.Primitive.Style.Hidden.html) - show or hide the primitive
+  * [`fill`](Scenic.Primitive.Style.Fill.html) - fill in the area of the primitive
+  * [`stroke`](Scenic.Primitive.Style.Stroke.html) - stroke the outline of the primitive. In this case, only the curvy part.
+  * [`join`](Scenic.Primitive.Style.Join.html) - control how segments are joined.
+  * [`miter_limit`](Scenic.Primitive.Style.MiterLimit.html) - control how segments are joined.
+  """
 
   use Scenic.Primitive
 
-  @styles [:hidden, :fill, :stroke]
+  @styles [:hidden, :fill, :stroke, :join, :miter_limit]
 
   # ============================================================================
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be: {width, height}
@@ -22,6 +42,7 @@ defmodule Scenic.Primitive.Rectangle do
     """
 
   # --------------------------------------------------------
+  @doc false
   def verify({width, height} = data) when is_number(width) and is_number(height) do
     {:ok, data}
   end
@@ -29,6 +50,9 @@ defmodule Scenic.Primitive.Rectangle do
   def verify(_), do: :invalid_data
 
   # ============================================================================
+  @doc """
+  Returns a list of styles recognized by this primitive.
+  """
   @spec valid_styles() :: [:fill | :hidden | :stroke, ...]
   def valid_styles(), do: @styles
 
@@ -36,6 +60,10 @@ defmodule Scenic.Primitive.Rectangle do
   def default_pin(data), do: centroid(data)
 
   # --------------------------------------------------------
+  @doc """
+  Returns the centroid of the rectangle. This is used as the default pin when applying
+  rotate or scale transforms.
+  """
   def centroid(data)
 
   def centroid({width, height}) do