+++ lib/scenic/primitive/sector.ex
@@ -4,7 +4,27 @@
 #
 
 defmodule Scenic.Primitive.Sector do
-  @moduledoc false
+  @moduledoc """
+  Draw an sector on the screen.
+
+  An sector is a shape that looks like a piece of pie.
+
+  ## Data
+
+  `{radius, start, finish}`
+
+  The data for a sector is a three-tuple.
+  * `radius` - the radius of the arc
+  * `start` - the starting angle in radians
+  * `finish` - end ending angle in radians
+
+  ## Styles
+
+  This primitive recognizes the following styles
+  * [`hidden`](Scenic.Primitive.Style.Hidden.html) - show or hide the primitive
+  * [`fill`](Scenic.Primitive.Style.Fill.html) - fill in the area of the primitive
+  * [`stroke`](Scenic.Primitive.Style.Stroke.html) - stroke the outline of the primitive. In this case, only the curvy part.
+  """
 
   use Scenic.Primitive
 
@@ -16,6 +36,7 @@ defmodule Scenic.Primitive.Sector do
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be: {radius, start_angle, end_angle}
@@ -24,6 +45,7 @@ defmodule Scenic.Primitive.Sector do
     """
 
   # --------------------------------------------------------
+  @doc false
   def verify(data) do
     normalize(data)
     {:ok, data}
@@ -32,12 +54,16 @@ defmodule Scenic.Primitive.Sector do
   end
 
   # --------------------------------------------------------
+  @doc false
   @spec normalize({number(), number(), number()}) :: {number(), number(), number()}
   def normalize({radius, start, finish} = data)
       when is_number(start) and is_number(finish) and is_number(radius),
       do: data
 
   # ============================================================================
+  @doc """
+  Returns a list of styles recognized by this primitive.
+  """
   @spec valid_styles() :: [:fill | :hidden | :stroke, ...]
   def valid_styles(), do: @styles