+++ lib/scenic/primitive/style/hidden.ex
@@ -4,7 +4,22 @@
 #
 
 defmodule Scenic.Primitive.Style.Hidden do
-  @moduledoc false
+  @moduledoc """
+  Flags whether or not to draw a primitive.
+
+  Example:
+
+      graph
+      |> rectangle({100, 200}, hidden: true)
+
+  ## Data
+  * `true` - "Hide" the primitive. Drawing is skipped.
+  * `false` - "Show" the primitive. Drawing is run.
+
+  Note: setting `hidden: true` on a group will hide all the primitives in
+  the group. This is very efficient as it simply skips drawing the group
+  and everything in it.
+  """
 
   use Scenic.Primitive.Style
 
@@ -12,6 +27,7 @@ defmodule Scenic.Primitive.Style.Hidden do
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be either true or false
@@ -20,6 +36,7 @@ defmodule Scenic.Primitive.Style.Hidden do
     """
 
   # --------------------------------------------------------
+  @doc false
   def verify(true), do: true
   def verify(false), do: true
   def verify(_), do: false