+++ lib/scenic/primitive/style/miter_limit.ex
@@ -4,7 +4,21 @@
 #
 
 defmodule Scenic.Primitive.Style.MiterLimit do
-  @moduledoc false
+  @moduledoc """
+  Automatically miter joints if they are too sharp.
+
+  Example:
+
+      graph
+      |> triangle( {{0,40},{40,40},{40,0}}
+        miter_limit: 2,
+        stroke: {2, :green}
+      )
+
+  ## Data
+
+  A number greater than zero.
+  """
 
   use Scenic.Primitive.Style
 
@@ -12,6 +26,7 @@ defmodule Scenic.Primitive.Style.MiterLimit do
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be a number greater than 0
@@ -20,8 +35,7 @@ defmodule Scenic.Primitive.Style.MiterLimit do
     """
 
   # --------------------------------------------------------
-  # named color
-
+  @doc false
   def verify(stroke) do
     try do
       normalize(stroke)
@@ -32,5 +46,6 @@ defmodule Scenic.Primitive.Style.MiterLimit do
   end
 
   # --------------------------------------------------------
+  @doc false
   def normalize(limit) when is_number(limit) and limit > 0, do: limit
 end