+++ lib/scenic/primitive/triangle.ex
@@ -4,17 +4,38 @@
 #
 
 defmodule Scenic.Primitive.Triangle do
-  @moduledoc false
+  @moduledoc """
+  Draw a triangle on the screen.
+
+  ## Data
+
+  `{point_a, point_b, point_c}`
+
+  The data for a line is a tuple containing three points.
+  * `point_a` - position to start drawing from
+  * `point_b` - position to draw to
+  * `point_c` - position to draw to
+
+  ## Styles
+
+  This primitive recognizes the following styles
+  * [`hidden`](Scenic.Primitive.Style.Hidden.html) - show or hide the primitive
+  * [`fill`](Scenic.Primitive.Style.Fill.html) - fill in the area of the primitive
+  * [`stroke`](Scenic.Primitive.Style.Stroke.html) - stroke the outline of the primitive. In this case, only the curvy part.
+  * [`join`](Scenic.Primitive.Style.Join.html) - control how segments are joined.
+  * [`miter_limit`](Scenic.Primitive.Style.MiterLimit.html) - control how segments are joined.
+  """
 
   use Scenic.Primitive
   alias Scenic.Math
 
-  @styles [:hidden, :fill, :stroke]
+  @styles [:hidden, :fill, :stroke, :join, :miter_limit]
 
   # ===========================================================================
   # data verification and serialization
 
   # --------------------------------------------------------
+  @doc false
   def info(data),
     do: """
       #{IO.ANSI.red()}#{__MODULE__} data must be three points: {{x0,y0}, {x1,y1}, {x2,y2}}
@@ -22,6 +43,7 @@ defmodule Scenic.Primitive.Triangle do
       #{IO.ANSI.default_color()}
     """
 
+  @doc false
   def verify({{x0, y0}, {x1, y1}, {x2, y2}} = data)
       when is_number(x0) and is_number(y0) and is_number(x1) and is_number(y1) and is_number(x2) and
              is_number(y2),
@@ -30,6 +52,9 @@ defmodule Scenic.Primitive.Triangle do
   def verify(_), do: :invalid_data
 
   # ============================================================================
+  @doc """
+  Returns a list of styles recognized by this primitive.
+  """
   @spec valid_styles() :: [:fill | :hidden | :stroke, ...]
   def valid_styles(), do: @styles
 
@@ -37,6 +62,10 @@ defmodule Scenic.Primitive.Triangle do
   def default_pin(data), do: centroid(data)
 
   # --------------------------------------------------------
+  @doc """
+  Returns the centroid of the triangle. This is used as the default pin when applying
+  rotate or scale transforms.
+  """
   def centroid(data)
 
   def centroid({{x0, y0}, {x1, y1}, {x2, y2}}) do